Split Sifrah GetDescription patch
